backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
moviepy
pandas
numpy
pymupdf
Pillow
proglog
pdfplumber
edge-tts
matplotlib
asyncio
ftfy
